+++ engine/src/index.ts
@@ -1,0 +1,11 @@
+// engine/src/index.ts
+
+export { lex } from "./core/lexer";
+export { parse } from "./core/parser";
+
+export { buildModuleGraph } from "./project/grapher";
+export type {
+  ModuleGraph,
+  FileAst,
+  ImportNode
+} from "./project/module";