+++ src/index.ts
@@ -1,6 +1,0 @@
-// src/index.ts
-export function GetVersion(): string {
-    return "1.0.0";
-}
-
-console.log(GetVersion());