+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.1] - 2025-12-20
+### Added
+- Split the repository into 'engine' and 'vscode' projects.
+- The version number for the separate projects start at 0.0.1
+
 ## [1.0.0] - 2025-11-22
 ### Added
 - Initial release.